+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - \#440 java.lang.IllegalArgumentException: URL is malformed
 - \#391 Parameters window too big for screen (Try to Restore last width) 
 - \#443 [Intellij] Error while loading a workspace 
+- \#450 Show log of a selected build
 
 ## [0.13.16]
 - \#415 Unable to enter parameters